fix: resolve clippy warnings for CI
Allow too_many_arguments crate-wide (layout draw functions are
inherently parameter-heavy), fix unnecessary casts, remove unused
import, use div_ceil, and suppress dead_code on items reserved for
future layouts.

+++ src/layout/classic.rs
@@ -22,7 +22,7 @@ impl super::Layout for ClassicLayout {
         let font_size = 6.0;
 
         let mut month_to_draw = config.month;
-        let mut year_to_draw = config.year as i32;
+        let mut year_to_draw = config.year;
 
         // First month: rotated 180°
         canvas.with_save(|c| {
